x11/py-sip6: Note that both newer versions are troubled
Both 6.11.1 and 6.12.0 cause the qgis build to fail, in a way that
suggests that upstream has made API/ABI breaks (hard to say if this is
P or B, but that's not the point).  I don't want to rush dealing with
this just before freeze, and will be trying to resolve the problems
for after the branch.

(If you want to test, updates to either are merely changing the
version/distinfo and building; there are not even PLIST changes.)