ghostscript-agpl: Work around illumos bug 17491.
It appears that ghostscript-agpl is deliberately using the system inclusion
guard namespace to do something clever so I don't want to patch this for every
OS, and it's unclear whether this is an OS bug or not, which is why it's a
subst hack limited to SunOS for now.  Fixes GCC 14 build.