mk/gcc.mk: don't pass "$" characters through a .for loop
Passing the expression "${.TARGET:H}" through the :S modifier relies on
the expression being preserved. This only works because the expression
is evaluated outside the context of a target.

Passing the same expression through the .for loop relies on undocumented
behavior of the .for loop, in that a "$" of an expression is not
escaped but preserves the expression.