wxGTK32: updated to 3.2.8.1
3.2.8.1

This is a hot fix release fixing a couple of regressions in 3.2.8 related to
ELF symbol versions, as the changes improving their handling done in 3.2.8
broke the build when using custom library flavour (i.e. --with-flavour
configure option) and also building using LTO with clang.

There are no other changes in this release compared to 3.2.8, so please refer
to this release announcement for more details about the changes since 3.2.7 and
more links.