mpg123: updated to 1.33.1
1.33.1

This has some little fixes, but mainly the fix for a regression since 1.31 is of interest: Repeated signals for terminal control were not handled, causing mpg123 to end unexpectedly on the second SIGUSR1, for example. The flaky semantics of signal() have bitten us. These remaining calls have been moved to sigaction() now like others. Full list of changes for 1.33.1 follows.

* INSTALL updated with hints for Windows, mainly.
* Finally formatted README. Maybe do it in Markdown sometime, as that's the fashion.
* build:
  - The ports/cmake only installs manpages for BUILD_PROGRAMS now.
  - The configure gives better hint if pkg-config was missing during generation (bug 378).
* mpg123:
  - Replace usage of signal() in terminal code with our sigaction() wrapper to fix repeated handling for --sigusr1 and --sigusr2, which got subtly broken on Linux + glibc by a feature test macro change in mpg123 1.31.