imath: update to 3.2.0.
## Version 3.2.0 (August 8, 2025)

Minor release with several bug and build fixes:

* Fix undefined behavior in the `Vec*` class `operator[]`
([#449](https://github.com/AcademySoftwareFoundation/Imath/pulls/449)).
* Fix call to `_mm_cvtps_ph` in `half.h`, Windows build failure

Build changes/improvements:
* Updated and simplified SO versioning policy: this release adopts a
  policy of appending the `MAJOR.MINOR.PATCH` software release name to
  the `SONAME` to form the real name of the shared library.
* Add `uninstall` target
* Add `iOS` cross-compilation support
* Add support for HIP (AMD's Heterogeneous-compute Interface for
  Portability)
* Fix handling of `/EHsc` flag
* Required cmake version is now 3.14

Also:
* Source files have been reformatted via `clang-format`.

This release also marks the introduction of an alternate boost-free
implementation of the Imath python module using pybind11. The
`pybindimath` module is an incomplete work-in-progress and is not
included in the installation. Activate the build via
`-DPYBIND11=ON`.