jjui: update to 0.9.1.
v0.9.1

This was supposed to be a maintenance release but it is packed with
new features thanks to all the contributors!

Phew, another release which was supposed to be a minor maintenance
release but to be honest, I didn't much other than implementing
flash messages and lane tracing.

New Features

Flash messages
Fuzzy file search
Ace jump
(experimental) Lane tracing
Restore from evolog
Preview window at the bottom of the UI
Force apply
Log batching